merkle: extract to new repository
