+++ Gopkg.lock
@@ -395,16 +395,18 @@
   revision = "aafc9e6bc7b7bb53ddaa75a5ef49a17d6e654be5"
 
 [[projects]]
-  name = "github.com/stretchr/objx"
-  packages = ["."]
-  revision = "facf9a85c22f48d2f52f2380e4efce1768749a89"
-  version = "v0.1"
+  branch = "master"
+  name = "github.com/stratumn/merkle"
+  packages = [
+    ".",
+    "types"
+  ]
+  revision = "b637554c3514b4b1a4a6de1eeb4ce0c1f2b91c37"
 
 [[projects]]
   name = "github.com/stretchr/testify"
   packages = [
     "assert",
-    "mock",
     "require"
   ]
   revision = "12b6f73e6084dad08a7c6e575284b177ecafbc71"
@@ -695,6 +697,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "4a504c3f995dc828b1f2c0b7893934924282956a3b51cb6321896d960185812e"
+  inputs-digest = "c13c05c83786528b68ff177233a9f5c671ad8e8915f59522814fcc7c4ba7e168"
   solver-name = "gps-cdcl"
   solver-version = 1
