--- merkle/main_test.go
@@ -1,35 +1,0 @@
-// Copyright 2017 Stratumn SAS. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package merkle_test
-
-import (
-	"flag"
-	"fmt"
-	"math/rand"
-	"os"
-	"testing"
-	"time"
-
-	"github.com/stratumn/go-indigocore/merkle/treetestcases"
-)
-
-func TestMain(m *testing.M) {
-	seed := int64(time.Now().Nanosecond())
-	fmt.Printf("using seed %d\n", seed)
-	rand.Seed(seed)
-	treetestcases.LoadFixtures("testdata")
-	flag.Parse()
-	os.Exit(m.Run())
-}
