--- merkle/testdata/path-a-0.json
@@ -1,1 +1,0 @@
-[]
