--- merkle/testdata/path-abc-1.json
@@ -1,12 +1,0 @@
-[
-  {
-    "left": "ca978112ca1bbdcafac231b39a23dc4da786eff8147c4e72b9807785afee48bb",
-    "right": "3e23e8160039594a33894f6564e1b1348bbd7a0088d42c4acb73eeaed59c009d",
-    "parent": "e5a01fee14e0ed5c48714f22180f25ad8365b53f9779f79dc4a3d7e93963f94a"
-  },
-  {
-    "left": "e5a01fee14e0ed5c48714f22180f25ad8365b53f9779f79dc4a3d7e93963f94a",
-    "right": "2e7d2c03a9507ae265ecf5b5356885a53393a2029d241394997265a1a25aefc6",
-    "parent": "7075152d03a5cd92104887b476862778ec0c87be5c2fa1c0a90f87c49fad6eff"
-  }
-]
