--- merkle/testdata/path-invalid-1.json
@@ -1,12 +1,0 @@
-[
-  {
-    "left": "2e7d2c03a9507ae265ecf5b5356885a53393a2029d241394997265a1a25aefc6",
-    "right": "18ac3e7343f016890c510e93f935261169d9e3f565436429830faf0934f4f8e4",
-    "parent": "bffe0b34dba16bc6fac17c08bac55d676cded5a4ade41fe2c9924a5dde8f3e5b"
-  },
-  {
-    "left": "14ede5e8e97ad9372327728f5099b95604a39593cac3bd38a343ad76205213e7",
-    "right": "3f79bb7b435b05321651daefd374cdc681dc06faa65e374e38337b88ca046dea",
-    "parent": "d71f8983ad4ee170f8129f1ebcdd7440be7798d8e1c80420bf11f1eced610dba"
-  }
-]
