--- merkle/util_test.go
@@ -1,33 +1,0 @@
-// Copyright 2017 Stratumn SAS. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package merkle_test
-
-import (
-	"encoding/json"
-	"io/ioutil"
-
-	"github.com/stratumn/go-indigocore/types"
-)
-
-func loadPath(filename string, path *types.Path) error {
-	data, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return err
-	}
-	if err = json.Unmarshal(data, path); err != nil {
-		return err
-	}
-	return nil
-}
