--- types/testdata/path-abc-2.json
@@ -1,7 +1,0 @@
-[
-  {
-    "left": "e5a01fee14e0ed5c48714f22180f25ad8365b53f9779f79dc4a3d7e93963f94a",
-    "right": "2e7d2c03a9507ae265ecf5b5356885a53393a2029d241394997265a1a25aefc6",
-    "parent": "7075152d03a5cd92104887b476862778ec0c87be5c2fa1c0a90f87c49fad6eff"
-  }
-]
