--- types/testdata/path-abcd-3.json
@@ -1,12 +1,0 @@
-[
-  {
-    "left": "2e7d2c03a9507ae265ecf5b5356885a53393a2029d241394997265a1a25aefc6",
-    "right": "18ac3e7343f016890c510e93f935261169d9e3f565436429830faf0934f4f8e4",
-    "parent": "bffe0b34dba16bc6fac17c08bac55d676cded5a4ade41fe2c9924a5dde8f3e5b"
-  },
-  {
-    "left": "e5a01fee14e0ed5c48714f22180f25ad8365b53f9779f79dc4a3d7e93963f94a",
-    "right": "bffe0b34dba16bc6fac17c08bac55d676cded5a4ade41fe2c9924a5dde8f3e5b",
-    "parent": "14ede5e8e97ad9372327728f5099b95604a39593cac3bd38a343ad76205213e7"
-  }
-]
