+++ types/types_test.go
@@ -565,3 +565,51 @@ func TestReversedBytes32Reverse(t *testing.T) {
 		}
 	}
 }
+
+func TestTransactionIDString(t *testing.T) {
+	str := "8353334c6e4911e6ad927bd17dea491a"
+	buf, _ := hex.DecodeString(str)
+	txid := types.TransactionID(buf)
+
+	if got, want := txid.String(), str; got != want {
+		t.Errorf("txid.String() = %q want %q", got, want)
+	}
+}
+
+func TestTransactionMarshalJSON(t *testing.T) {
+	str := "8353334c6e4911e6ad927bd17dea491a"
+	buf, _ := hex.DecodeString(str)
+	txid := types.TransactionID(buf)
+	marshalled, err := json.Marshal(txid)
+	if err != nil {
+		t.Fatalf("json.Marshal(): err: %s", err)
+	}
+
+	if got, want := string(marshalled), fmt.Sprintf(`"%s"`, str); got != want {
+		t.Errorf("txid.MarshalJSON() = %q want %q", got, want)
+	}
+}
+
+func TestTransactionUnmarshalJSON(t *testing.T) {
+	str := "8353334c6e4911e6ad927bd17dea491a"
+	marshalled := fmt.Sprintf(`"%s"`, str)
+	var txid types.TransactionID
+	err := json.Unmarshal([]byte(marshalled), &txid)
+	if err != nil {
+		t.Fatalf("json.Unmarshal(): err: %s", err)
+	}
+
+	if got, want := txid.String(), str; got != want {
+		t.Errorf("txid.UnmarshalJSON() = %q want %q", got, want)
+	}
+}
+
+func TestTransactionUnmarshalJSON_invalid(t *testing.T) {
+	str := "azertyu"
+	marshalled := fmt.Sprintf(`"%s"`, str)
+	var txid types.TransactionID
+	err := json.Unmarshal([]byte(marshalled), &txid)
+	if err == nil {
+		t.Error("json.Unmarshal(): err = nil want Error")
+	}
+}
